requirements.txt
google-genai
python-dotenv
pandas
pypdf

langchain-core
langchain-text-splitters
langchain-postgres
sqlalchemy
asyncpg
